Market/requirements.txt
aiogram
aiohttp
asyncpg
python-dotenv
openpyxl